+++ .env
@@ -1,0 +1,6 @@
+VITE_FIREBASE_API_KEY='your_api_key_here'
+VITE_FIREBASE_AUTH_DOMAIN='your_auth_domain_here'
+VITE_FIREBASE_PROJECT_ID='your_project_id_here'
+VITE_FIREBASE_STORAGE_BUCKET='your_storage_bucket_here'
+VITE_FIREBASE_MESSAGING_SENDER_ID='your_messaging_sender_id_here'
+VITE_FIREBASE_APP_ID='your_app_id_here'
