+++ .gitignore
@@ -12,6 +12,9 @@ dist
 dist-ssr
 *.local
 
+# production
+/build
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
