+++ package.json
@@ -15,6 +15,7 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@mui/material": "^6.4.1",
+    "env-cmd": "^10.1.0",
     "firebase": "^11.2.0",
     "i18next": "^24.2.2",
     "i18next-browser-languagedetector": "^8.0.2",
