+++ src/pages/Landing/sections/Footer/Footer.scss
@@ -1,0 +1,21 @@
+@use '../../../../theme/styles.scss' as *;
+@use '../../../../theme/colors.scss' as *;
+
+.footer {
+  width: 100%;
+  min-width: 400px;
+  height: 10vh;
+    
+  &.light{
+    background: $gradient-to-top;
+    color: white
+  }
+  &.dark{
+    background: $gradient-top-bottom;
+    color: white;
+  }
+  p {
+    font-weight: 500;
+    font-size: 1em;
+  }
+}
