+++ src/pages/Landing/sections/Navbar/Navbar.scss
@@ -1,0 +1,37 @@
+@use '../../../../theme/colors.scss' as *;
+
+.navbar {
+  height: 60px;
+  display: flex;
+  min-width: 300px;
+  position: fixed;
+  padding: 0 50px;
+  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.1);
+  .name {
+    font-weight: 600;
+    font-size: 20px;
+    color: $primary-color;
+    cursor: pointer;
+  }
+  .navbar-item {
+    padding: 15px;
+    text-decoration: none;
+    transition: transform 0.3s ease;
+    &.theme{
+      transform: translateY(0px);
+      &.dark{
+        filter: invert(1);
+      }
+      &:hover{
+        transform: translateY(-2px);
+      }
+    }
+    &:hover {
+      color: $primary-color;
+      cursor: pointer;
+    }
+  }
+ 
+}
+ 
+
