+++ src/pages/Landing/sections/Profile/Profile.scss
@@ -1,0 +1,77 @@
+@use '../../../../theme/styles.scss' as *;
+@use '../../../../theme/colors.scss' as *;
+
+.profile {
+  padding-top: 50px;
+  padding-bottom: 10px;
+  margin-top: 50px;
+  min-width: 400px;
+  flex-direction: column;
+  .info {
+    gap: 10%;
+    .me {
+      width: 30%;
+      p:nth-child(1) {
+        animation: scale-fade-in 0.2s ease-out forwards;
+        @include gradient(black);
+        &.dark {
+          @include gradient($secondary-color-light);
+          -webkit-background-clip: text;
+          background-clip: text;
+          color: transparent;
+        }
+        -webkit-background-clip: text;
+        background-clip: text;
+        color: transparent;
+        font-size: 45px;
+        font-weight: bold;
+        line-height: 1;
+      }
+      p:nth-child(2) {
+        animation: scale-fade-in 0.3s ease-out forwards;
+        font-size: 20px;
+        font-weight: 500;
+        line-height: 1;
+      }
+      p:nth-child(3) {
+        animation: scale-fade-in 0.4s ease-out forwards;
+        font-size: 15px;
+        line-height: 1.5;
+      }
+      span {
+        color: $primary-color;
+      }
+    }
+
+    .photo {
+      animation: scale-fade-in 0.5s ease-out forwards;
+      position: relative;
+      height: 250px;
+      width: 250px;
+      border-radius: 100%;
+      background: $gradient-bottom-topRight;
+      box-shadow: 0 4px 6px rgba(0, 0, 0, 0.19);
+      img {
+        width: 100%;
+        height: 100%;
+        border-radius: inherit;
+      }
+    }
+    .icon-download {
+      position: absolute;
+      bottom: 10%;
+      left: 0%;
+    }
+  }
+
+  @media (max-width: 750px) {
+    .info {
+      display: flex;
+      flex-direction: column-reverse;
+      .me {
+        width: 70%;
+        text-align: center;
+      }
+    }
+  }
+}
