+++ src/pages/Loading/Loading.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './Loading.scss';
+
+const LoadingScreen: React.FC = () => {
+  return (
+    <div className="loading-screen">
+      <div className="spinner"></div>
+    </div>
+  );
+};
+
+export default LoadingScreen;
