--- src/pages/sections/About/About.scss
@@ -1,47 +1,0 @@
-@use '../../../theme/colors.scss' as *;
-@use '../../../theme/styles.scss' as *;
-
-.about-me {
-  width: 100%;
-  min-width: 400px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 50px;
-  background-color: $default-color-white;
-  @include background-transition;
-  
-  &.dark {
-    background-color: $tertiary-background-color;
-    .card {
-      background-color: $default-color-card-dark;
-    }
-  }
-  .me {
-    @include title-style;
-  }
-
-  .card {
-    width: 100%;
-    max-width: 650px;
-    min-width: 400px;
-    padding: 24px;
-    border-radius: 8px;
-    background-color: $default-color-card;
-    transition: transform 0.3s ease-in-out;
-    font-weight: 400;
-    font-size: 0.8em;
-    text-decoration: none;
-    text-align: justify;
-  }
-
-  @media (max-width: 720px) {
-    .card {
-      box-sizing: content-box;
-      min-width: 100px;
-      width: 80%;
-      text-align: center;
-    }
-  }
-}
