--- src/pages/sections/Contact/Contact.scss
@@ -1,19 +1,0 @@
-@use '../../../theme/colors.scss' as *;
-@use '../../../theme/styles.scss' as *;
-
-.contact {
-  width: 100%;
-  min-width: 400px;
-  padding-bottom: 64px;
-  gap: 34px;
-  p {
-    @include title-style;
-  }
-
-  &.dark {
-    background-color: $quaternary-background-color;
-  }
-  .card-c {
-    gap: 124px;
-  }
-}
