--- src/pages/sections/Contact/sections/Rombo/Rombo.scss
@@ -1,36 +1,0 @@
-@use '../../../../../theme/colors.scss' as *;
-.rombo {
-  width: 70px;
-  height: 70px;
-  position: relative;
-  border-radius: 100px;
-  transform: rotate(45deg);
-  text-decoration: none;
-  transition: transform 0.3s ease;
-
-  img {
-    width: 80%;
-    height: 80%;
-    object-fit: cover;
-    transform: rotate(-45deg);
-  }
-
-  &:hover,
-  &:active,
-  &:focus-visible {
-    transform: rotate(405deg) scale(1.2);
-    outline: none;
-    &.github {
-      filter: drop-shadow(0 0 20px rgba(0, 0, 0, 0.5));
-      &.dark {
-        filter: drop-shadow(0 0 20px rgba(255, 255, 255, 0.5));
-      }
-    }
-    &.whatsapp {
-      filter: drop-shadow(0 0 20px rgba(0, 255, 0, 0.5));
-    }
-    &.linkedin {
-      filter: drop-shadow(0 0 20px rgba(0, 166, 255, 0.5));
-    }
-  }
-}
