--- src/pages/sections/Profile/components/ButtonDownload/ButtonDownload.scss
@@ -1,64 +1,0 @@
-@use '../../../../../theme/colors.scss' as *;
-@use '../../../../../theme/styles.scss' as *;
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    visibility: hidden;
-  }
-  to {
-    opacity: 1;
-    visibility: visible;
-  }
-}
-
-.button-download {
-  border-radius: 100%;
-  height: 50px;
-  width: 25px;
-  padding: 0 12px;
-  background: $gradient-to-top;
-  text-decoration: none;
-  color: white;
-  cursor: pointer;
-  transition: all 0.3s ease;
-
-  overflow: hidden;
-
-  .download-text {
-    opacity: 0;
-    visibility: hidden;
-
-    transition:
-      opacity 0.3s ease,
-      visibility 0.3s ease;
-    margin: 0;
-  }
-
-  &:hover,
-  &:focus-visible {
-    transform: scale(1.1);
-    width: 120px;
-    transition:
-      width 0.3s ease,
-      transform 0.3s ease;
-    box-shadow: $box-shadow-card;
-    border-radius: 8px;
-    background: $primary-color;
-    .download-text {
-      opacity: 1;
-      visibility: visible;
-      margin-left: 8px;
-      animation: fadeIn 0.3s ease forwards;
-    }
-  }
-
-  &:active {
-    transform: scale(0.9);
-  }
-
-  .icon {
-    height: 24px !important;
-    width: 24px !important;
-  }
-}
