--- src/pages/sections/Projects/sections/ProjectCard.scss
@@ -1,85 +1,0 @@
-@use '../../../../theme/styles.scss' as *;
-@use '../../../../theme/colors.scss' as *;
-
-.project-card {
-  width: 330px;
-  height: 400px;
-  background-color: white;
-  border-radius: 4px;
-  @media screen and (max-width: 700px) {
-    height: 300px;
-  }
-  &.dark {
-    box-shadow: $box-shadow-card-medium;
-    background-color: $default-color-card-dark;
-  }
-
-  .image {
-    margin: 8px 8px 0 8px;
-    height: 60%;
-    object-fit: cover;
-  }
-
-  .card-content {
-    width: 100%;
-    height: 100%;
-    padding: 8px 8px 12px 8px;
-    box-sizing: border-box;
-
-    .title {
-      color: $primary-color;
-      font-weight: bold;
-    }
-
-    .description {
-      color: rgb(173, 173, 173);
-      font-size: 0.8em;
-      display: -webkit-box;
-      -webkit-line-clamp: 3;
-      -webkit-box-orient: vertical;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      word-break: break-word;
-      white-space: normal;
-    }
-    .actions {
-      margin-top: 4px;
-      &.dark {
-        .demo {
-          color: $default-color-white;
-        }
-        .details {
-          background-color: $tertiary-background-color;
-        }
-      }
-      .demo {
-        background: none;
-        border: none;
-        color: rgb(53, 53, 53);
-        padding: 8px;
-        transition:
-          color 0.3s ease,
-          background-color 0.3s ease,
-          border-radius 0.3s ease;
-        &:hover {
-          cursor: pointer;
-          border-radius: 4px;
-          background-color: rgba(180, 180, 180, 0.103);
-        }
-      }
-      .details {
-        padding: 8px;
-        border-radius: 4px;
-        border: none;
-        background-color: $primary-color;
-        color: $default-color-white;
-        font-weight: 500;
-        cursor: pointer;
-        transition: transform 0.3s ease;
-        &:hover {
-          transform: scale(0.9);
-        }
-      }
-    }
-  }
-}
