+++ src/shared/constants/firebase.ts
@@ -1,10 +1,11 @@
 export const COLLECTION_NAME = 'projects';
 
 export const FIREBASE_CONFIG = {
-  apiKey: 'AIzaSyBMbC2s8Od_rixh0POAZOTIv5VvIj5k1JU',
-  authDomain: 'react-portfolioweb.firebaseapp.com',
-  projectId: 'react-portfolioweb',
-  storageBucket: 'react-portfolioweb.firebasestorage.app',
-  messagingSenderId: '466783937622',
-  appId: '1:466783937622:web:78a03f1c3bd03bc9ffc7f7'
+  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
+  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
+  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_FIREBASE_APP_ID
 };
+
