+++ src/shared/constants/user.tsx
@@ -1,0 +1,1 @@
+export const USER_NAME = 'Richard Morales';
