+++ src/shared/translations/i18n.ts
@@ -15,6 +15,9 @@ use(Backend)
     interpolation: {
       escapeValue: false
     },
+    react: {
+      useSuspense: true
+    },
     resources: {
       es: {
         translation: esFile
