+++ tsconfig.json
@@ -15,6 +15,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
+    "types": ["vite/client"],
   },
   "include": ["src", "svg.d.ts"],
   "exclude": ["node_modules", "dist"]
